intellibin/requirements.txt
djangorestframework
python-dotenv
psycopg
# psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
django-otp
pyotp